Module B000_conv.Op

B000.Op converters

Stringifiers

val status_to_string : B000.Op.status -> string

status_to_string s is s as a string.

val notify_kind_to_string : B000.Op.Notify.kind -> string

notify_kind_to_string k is k as a string.

Formatters

val pp_file_read : B0_std.Fpath.t B0_std.Fmt.t

pp_file_read formats a read file.

val pp_file_write : B0_std.Fpath.t B0_std.Fmt.t

pp_file_write formats a written file.

val pp_line : B000.Op.t B0_std.Fmt.t

pp_line formats a build operation on a single line.

val pp_line_and_ui : B000.Op.t B0_std.Fmt.t

pp_line_and_ui formats like pp_line but also a potential feedback spawn UI and/or failure error messages.

val pp_ui : sep:unit B0_std.Fmt.t -> op_howto:B000.Op.t B0_std.Fmt.t -> B000.Op.t B0_std.Fmt.t

pp_ui formats notifications, failed operations and spawn operations with a UI. In case something is printed sep is added at the end. op_howto should be command fragment to obtain more information about operation failures.

val pp : B000.Op.t B0_std.Fmt.t

pp formats a build operation with full details.

val bincode : B000.Op.t B0_std.Bincode.t

bincode binary codes an operation.

Aggregate errors

val pp_aggregate_error : ?sep:unit B0_std.Fmt.t -> ?read_howto:B0_std.Fpath.t B0_std.Fmt.t -> ?write_howto:B0_std.Fpath.t B0_std.Fmt.t -> unit -> B000.Op.aggregate_error B0_std.Fmt.t

pp_aggregate_error ~read_howto ~write_howto formats a memo error followed by sep iff somethings is printed (defaults to B0_std.Fmt.flush_nl). The errors are formatted as follows:

B000.Op.aggregate_error.Failures formats B0_std.Fmt.nop.

B000.Op.aggregate_error.Never_became_ready formats each file prefixing it with read_howto.

B000.Op.aggregate_error.Cycle, formats the operations of the cycle. by writing the files that form the cycle and prefixing them with write_howto.